# docs/swagger.yaml
definitions:
...
  fitness-tracker-backend_workout_handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.muscleGroupRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  handler.profileStatsResponse:
    properties:
      ai_insight:
        type: string
      last_training_at:
        type: string
      total_exercise_sessions:
        type: integer
      total_workouts:
        type: integer
      trainings_this_week:
        type: integer
    type: object
...
  models.MuscleGroup:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for the Fitness Tracker backend service.
  title: Fitness Tracker API
  version: "1.0"
paths:
...
  /muscle-groups:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MuscleGroup'
            type: array
      security:
      - BearerAuth: []
      summary: List muscle groups
      tags:
      - muscle-groups
    post:
      consumes:
      - application/json
      parameters:
      - description: Muscle group
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.muscleGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.MuscleGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create muscle group
      tags:
      - muscle-groups
  /muscle-groups/{id}:
    delete:
      parameters:
      - description: MuscleGroup ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete muscle group
      tags:
      - muscle-groups
    get:
      parameters:
      - description: MuscleGroup ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MuscleGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get muscle group by ID
      tags:
      - muscle-groups
    put:
      consumes:
      - application/json
      parameters:
      - description: MuscleGroup ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.muscleGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MuscleGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Update muscle group
      tags:
      - muscle-groups
...
  /profile/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.profileStatsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get current user's workout stats
      tags:
      - profile
...
securityDefinitions:
  BearerAuth:
    description: 'Provide your JWT with the "Bearer " prefix. Example: "Bearer {token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"